feat: add a help script to run the test over RESTful APIs
Modifications:
1. add a help script to run the test over RESTful APIs
2. check upload_dir in the runner before packing resouces

Tests:
1. demo test passes
2. wifi basic test passes

Reviewers:
wei.li

+++ webrobot/task-runner/runner.py
@@ -90,10 +90,11 @@ def run_task_for_endpoint(endpoint):
                             task.status = 'failed'
                         task.save()
 
-                        resource_dir_tmp = Path(get_config().UPLOAD_ROOT) / task.upload_dir
-                        if resource_dir_tmp != '' and os.path.exists(resource_dir_tmp):
-                            make_tarfile(str(result_dir / 'resource.tar.gz'), resource_dir_tmp)
-                            shutil.rmtree(resource_dir_tmp)
+                        if task.upload_dir:
+                            resource_dir_tmp = Path(get_config().UPLOAD_ROOT) / task.upload_dir
+                            if resource_dir_tmp != '' and os.path.exists(resource_dir_tmp):
+                                make_tarfile(str(result_dir / 'resource.tar.gz'), resource_dir_tmp)
+                                shutil.rmtree(resource_dir_tmp)
 
                         send_email(task)
